+++ src/zarzadzanieekwipunkiem.h
@@ -15,7 +15,7 @@ void dezaktywujBonusy(Przedmiot* przedmiot, Gracz* gracz);
 bool czyDozwolony(Przedmiot* przedmiot, Gracz* gracz);
 bool czyBrakOgraniczenia(Przedmiot* przedmiot, int indeks);
 
-QString dzialanie(quint8 bonus, QString dzialanie);
+QString dzialanie(int bonus, QString dzialanie);
 void wygenerujOpis(Przedmiot* rzecz, Gracz* gracz, QTextBrowser* miejsce);
 
 bool czyZalozony(Przedmiot* rzecz, Gracz* gracz);
